requirements.txt
attrdict==2.0.1
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
dominate==2.6.0
Flask==2.1.3
Flask-Cors==3.0.10
idna==3.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonpatch==1.32
jsonpointer==2.3
MarkupSafe==2.1.1
numpy==1.23.1
opencv-python==4.6.0.66
Pillow==9.2.0
PyYAML==6.0
pyzmq==23.2.0
requests==2.28.1
scipy==1.8.1
six==1.16.0
torch==1.12.0
torchfile==0.1.0
torchvision==0.13.0
tornado==6.2
typing_extensions==4.3.0
urllib3==1.26.10
visdom==0.1.8.9
websocket-client==1.3.3
Werkzeug==2.1.2
zipp==3.8.1
